move docs to settings renamed as user guide
