+++ client/packages/common/src/intl/locales/en/common.json
@@ -433,6 +433,7 @@
   "heading.description": "Description",
   "heading.details": "Details",
   "heading.devices": "Devices",
+  "heading.user-guide": "User Guide",
   "heading.dispensary-details": "Dispensary Details",
   "heading.download-catalogue-documents": "Download catalogue documents",
   "heading.download-cce-documents": "Download CCE documents",
@@ -584,6 +585,7 @@
   "label.click-to-sort": "Click to sort by ",
   "label.client": "Client",
   "label.clinician": "Clinician",
+  "label.user-guide": "Online User Guide (requires internet connection)",
   "label.code": "Code",
   "label.code-or-name": "Code or Name",
   "label.cold-consecutive": "Cold Consecutive",
