+++ client/packages/host/src/Admin/Settings.tsx
@@ -12,6 +12,7 @@ import {
   SunIcon,
   PrinterIcon,
   ListIcon,
+  BookIcon,
 } from '@common/icons';
 import { AppVersion } from '../components';
 import { SyncSettings } from './SyncSettings';
@@ -22,6 +23,7 @@ import { DisplaySettings } from './DisplaySettings';
 import { SettingsSection } from './SettingsSection';
 import { LabelPrinterSettings } from './LabelPrinterSettings';
 import { ConfigurationSettings } from './ConfigurationSettings';
+import { UserGuide } from './UserGuide';
 
 export const Settings: React.FC = () => {
   const { data: initStatus } = useInitialisationStatus();
@@ -77,6 +79,14 @@ export const Settings: React.FC = () => {
           <ConfigurationSettings />
         </SettingsSection>
       )}
+      <SettingsSection
+        Icon={BookIcon}
+        titleKey="heading.user-guide"
+        expanded={activeSection === 5}
+        onChange={toggleSection(5)}
+      >
+        <UserGuide />
+      </SettingsSection>
       <AppBarButtonsPortal>
         <AppVersion SiteInfo={<SiteInfo siteName={initStatus?.siteName} />} />
       </AppBarButtonsPortal>
