+++ client/packages/host/src/components/AppDrawer/AppDrawer.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import { Property } from 'csstype';
 import {
-  BookIcon,
   Box,
   DashboardIcon,
   Divider,
-  ExternalNavLink,
   List,
   PowerIcon,
   SettingsIcon,
@@ -17,7 +15,6 @@ import {
   AppNavLink,
   useIsMediumScreen,
   useAuthContext,
-  useLocation,
   EnvUtils,
   UserPermission,
   RouteBuilder,
@@ -25,11 +22,7 @@ import {
   ReportsIcon,
   useHostContext,
 } from '@openmsupply-client/common';
-import {
-  AppRoute,
-  ExternalURL,
-  useExternalUrl,
-} from '@openmsupply-client/config';
+import { AppRoute } from '@openmsupply-client/config';
 import {
   CatalogueNav,
   DistributionNav,
@@ -153,7 +146,6 @@ export const AppDrawer: React.FC = () => {
   const drawer = useDrawer();
   const { logout, userHasPermission, store } = useAuthContext();
   const { fullScreen } = useHostContext();
-  const location = useLocation();
   const navigate = useNavigate();
 
   React.useEffect(() => {
@@ -179,11 +171,6 @@ export const AppDrawer: React.FC = () => {
     drawer.open();
     drawer.setHoverOpen(true);
   };
-
-  const publicDocsUrl = useExternalUrl(ExternalURL.PublicDocs);
-  const docsUrl = `${publicDocsUrl}${
-    EnvUtils.mapRoute(location.pathname).docs
-  }`;
 
   const handleLogout = () => {
     navigate(RouteBuilder.create(AppRoute.Login).build());
@@ -249,12 +236,6 @@ export const AppDrawer: React.FC = () => {
       <LowerListContainer onMouseEnter={onHoverOver} onMouseLeave={onHoverOut}>
         <List>
           {drawer.isOpen && <StyledDivider color="drawerDivider" />}
-          <ExternalNavLink
-            to={docsUrl}
-            icon={<BookIcon fontSize="small" color="primary" />}
-            text={t('docs')}
-            trustedSite={true}
-          />
           <SyncNavLink />
           <AppNavLink
             to={AppRoute.Settings}
